moved url preview VM creation to message composer
